# docs/api/openapi-spec.yaml
openapi: 3.0.0
info:
  title: SmartWalletFX Crypto Data API
  version: 3.0.0
  description: |
    **Phase 3 Production Release** - Enterprise-grade cryptocurrency portfolio and DeFi data API with 
    comprehensive monitoring, advanced cost optimization, and multi-chain support.
    
    ## 🎯 Phase 3 Enhancements
    - **Security Hardening**: Enhanced rate limiting, input validation, and audit logging
    - **Performance Optimization**: 42% cost reduction, sub-200ms response times
    - **Build Improvements**: Automated testing, CI/CD pipeline, container optimization
    - **Monitoring Excellence**: 15+ Prometheus metrics, comprehensive alerting
    
    ## ✨ Key Features
    - **Multi-chain Support**: Ethereum, Polygon, Arbitrum, Optimism, Base, Solana
    - **DeFi Integration**: 15+ protocols (Aave V3, Compound V3, Uniswap V3, Curve, Yearn, Lido)
    - **NFT Portfolio**: Collection tracking, metadata enrichment, rarity analysis
    - **Intelligent Caching**: Multi-level (L1/L2/L3) with predictive warming
    - **Cost Optimization**: Smart batching, provider rotation, request deduplication
    - **Enterprise Observability**: Structured logging, distributed tracing, cost tracking
    
    ## 🔐 Authentication
    All endpoints require an API key in the `X-API-Key` header unless specified otherwise.
    
    ## 📊 Rate Limits
    - **Standard**: 1000 requests/minute
    - **Premium**: 5000 requests/minute
    - **Enterprise**: Custom limits available
    
  contact:
...
  license:
...
servers:
...
security:
  - ApiKeyAuth: []
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        API key for authentication. Obtain your API key from the SmartWalletFX dashboard.
        
        Example: `X-API-Key: swfx_1234567890abcdef`
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT bearer token for enhanced security (Enterprise only)
  schemas:
...
    SuccessResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          required:
            - data
          properties:
            data:
              type: object
              description: Response data payload
...
    PortfolioData:
      type: object
      required:
        - address
        - totalValueUSD
        - lastUpdated
      properties:
        address:
          type: string
          description: Wallet address
          example: "0x742d35Cc6634C0532925a3b8D60c7b35C4D2C14a"
        totalValueUSD:
          type: number
          format: decimal
          description: Total portfolio value in USD
          example: 12500.75
        netWorthUSD:
          type: number
          format: decimal
          description: Net worth (assets minus debts)
          example: 11800.50
        defi:
          $ref: '#/components/schemas/DeFiSummary'
        nfts:
          $ref: '#/components/schemas/NFTSummary'
        tokens:
          $ref: '#/components/schemas/TokenSummary'
        chainDistribution:
          type: array
          items:
            $ref: '#/components/schemas/ChainDistribution'
        riskMetrics:
          $ref: '#/components/schemas/RiskMetrics'
        lastUpdated:
          type: string
          format: date-time
          description: Last portfolio update timestamp
...
    NFTSummary:
      type: object
      properties:
        totalValueUSD:
          type: number
          format: decimal
          description: Total estimated NFT value
          example: 3500.00
        totalCollections:
          type: integer
          description: Number of collections owned
          example: 8
        totalNFTs:
          type: integer
          description: Total number of NFTs
          example: 45
        floorValueUSD:
          type: number
          format: decimal
          description: Total floor value
          example: 2800.00
...
    NFTCollection:
      type: object
      required:
        - contractAddress
        - name
        - tokenCount
        - chainId
      properties:
        contractAddress:
          type: string
          description: Collection contract address
          example: "0xBC4CA0EdA7647A8aB7C2061c2E118A18a936f13D"
        name:
          type: string
          description: Collection name
          example: "Bored Ape Yacht Club"
        symbol:
          type: string
          description: Collection symbol
          example: "BAYC"
        tokenCount:
          type: integer
          description: Number of NFTs owned in collection
          example: 2
        floorPriceUSD:
          type: number
          format: decimal
          description: Current floor price
          example: 15000.00
        totalValueUSD:
          type: number
          format: decimal
          description: Total estimated value
          example: 30000.00
        chainId:
          type: string
          description: Blockchain chain ID
          example: "1"
        tokens:
          type: array
          items:
            $ref: '#/components/schemas/NFTToken'
...
    ProtocolInfo:
      type: object
      required:
        - id
        - name
        - type
        - chains
      properties:
        id:
          type: string
          description: Protocol identifier
          example: "aave-v3"
        name:
          type: string
          description: Protocol name
          example: "Aave V3"
        type:
          type: string
          enum: [lending, dex, farming, staking, insurance, derivatives]
          description: Protocol category
        description:
          type: string
          description: Protocol description
        website:
          type: string
          format: uri
          description: Protocol website
        logoUri:
          type: string
          format: uri
          description: Protocol logo URL
        chains:
          type: array
          items:
            type: string
          description: Supported chain IDs
          example: ["1", "137", "42161"]
        tvl:
          type: number
          format: decimal
          description: Total Value Locked in USD
        isActive:
          type: boolean
          description: Whether protocol is currently active

    # Pagination
...
  parameters:
    AddressParam:
      name: address
      in: path
      required: true
      schema:
        type: string
        pattern: '^0x[a-fA-F0-9]{40}$|^[1-9A-HJ-NP-Za-km-z]{32,44}$'
      description: Wallet address (Ethereum format or Solana format)
      example: "0x742d35Cc6634C0532925a3b8D60c7b35C4D2C14a"
    ChainIdsParam:
      name: chains
      in: query
      required: false
      schema:
        type: string
      description: Comma-separated list of chain IDs
      example: "1,137,42161"
    IncludeParam:
      name: include
      in: query
      required: false
      schema:
        type: string
      description: Comma-separated list of data to include
      example: "defi,nfts,tokens"
    LimitParam:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Number of items to return
    PageParam:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Page number for pagination
    RefreshParam:
      name: refresh
      in: query
      required: false
      schema:
        type: boolean
        default: false
      description: Force refresh of cached data
  responses:
    BadRequest:
      description: Bad request - invalid parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: "INVALID_ADDRESS"
              message: "The provided address format is invalid"
              details:
                address: "invalid-address"
                expectedFormat: "0x followed by 40 hexadecimal characters"
            metadata:
              timestamp: "2024-01-15T10:30:00.000Z"
              requestId: "550e8400-e29b-41d4-a716-446655440000"
    Unauthorized:
      description: Unauthorized - invalid or missing API key
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: "UNAUTHORIZED"
              message: "Invalid or missing API key"
            metadata:
              timestamp: "2024-01-15T10:30:00.000Z"
              requestId: "550e8400-e29b-41d4-a716-446655440000"
    RateLimit:
      description: Rate limit exceeded
      headers:
        X-RateLimit-Limit:
          description: Request limit per window
          schema:
            type: integer
        X-RateLimit-Remaining:
          description: Remaining requests in window
          schema:
            type: integer
        X-RateLimit-Reset:
          description: Window reset time
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
paths:
...
  /api/portfolio/{address}:
    get:
      tags:
        - Portfolio
      summary: Get complete portfolio
      description: |
        Retrieves comprehensive portfolio data including DeFi positions, 
        NFTs, tokens, and risk metrics for a given address.
      parameters:
        - $ref: '#/components/parameters/AddressParam'
        - $ref: '#/components/parameters/ChainIdsParam'
        - $ref: '#/components/parameters/IncludeParam'
        - $ref: '#/components/parameters/RefreshParam'
      responses:
        '200':
          description: Portfolio data retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PortfolioData'
              example:
                success: true
                data:
                  address: "0x742d35Cc6634C0532925a3b8D60c7b35C4D2C14a"
                  totalValueUSD: 12500.75
                  netWorthUSD: 11800.50
                  defi:
                    totalValueUSD: 8500.25
                    positionCount: 12
                    protocolCount: 4
                  nfts:
                    totalValueUSD: 3500.00
                    totalCollections: 8
                    totalNFTs: 45
                  lastUpdated: "2024-01-15T10:30:00.000Z"
                metadata:
                  timestamp: "2024-01-15T10:30:00.000Z"
                  requestId: "550e8400-e29b-41d4-a716-446655440000"
                  duration: 150
                  cacheStatus: "hit"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimit'
        '500':
          $ref: '#/components/responses/InternalError'

  # DeFi Endpoints
...
  /api/defi/protocols:
    get:
      tags:
        - DeFi
      summary: List supported protocols
      description: Returns a list of all supported DeFi protocols with their details
      parameters:
        - name: chainId
          in: query
          required: false
          schema:
            type: string
          description: Filter protocols by chain ID
        - name: type
          in: query
          required: false
          schema:
            type: string
            enum: [lending, dex, farming, staking, insurance, derivatives]
          description: Filter protocols by type
      responses:
        '200':
          description: Protocols list retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          protocols:
                            type: array
                            items:
                              $ref: '#/components/schemas/ProtocolInfo'
                          totalCount:
                            type: integer

  # NFT Endpoints
  /api/nft/{address}:
    get:
      tags:
        - NFT
      summary: Get NFT portfolio
      description: |
        Retrieves all NFT collections and tokens owned by an address.
        Includes metadata, rarity information, and value estimates.
      parameters:
        - $ref: '#/components/parameters/AddressParam'
        - $ref: '#/components/parameters/ChainIdsParam'
        - name: includeMetadata
          in: query
          required: false
          schema:
            type: boolean
            default: true
          description: Include full NFT metadata
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/PageParam'
      responses:
        '200':
          description: NFT portfolio retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          collections:
                            type: array
                            items:
                              $ref: '#/components/schemas/NFTCollection'
                          summary:
                            $ref: '#/components/schemas/NFTSummary'
                          pagination:
                            $ref: '#/components/schemas/PaginationInfo'
...
  /api/solana/{address}:
    get:
      tags:
        - Solana
      summary: Get Solana portfolio
      description: |
        Retrieves Solana-specific portfolio data including SPL tokens,
        NFTs, and DeFi positions on Solana protocols.
      parameters:
        - name: address
          in: path
          required: true
          schema:
            type: string
            pattern: '^[1-9A-HJ-NP-Za-km-z]{32,44}$'
          description: Solana wallet address
          example: "9WKKVhUG8L8YHxKUpHqVYKKrE1M1hJ5PeG8Fw7tpFm3z"
        - name: include
          in: query
          required: false
          schema:
            type: string
          description: Data to include (tokens,nfts,defi,staking)
          example: "tokens,defi,staking"
      responses:
        '200':
          description: Solana portfolio retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          tokens:
                            type: array
                            items:
                              $ref: '#/components/schemas/TokenInfo'
                          nfts:
                            $ref: '#/components/schemas/NFTSummary'
                          defi:
                            $ref: '#/components/schemas/DeFiSummary'
                          staking:
                            type: object
                            properties:
                              totalStaked:
                                type: number
                                format: decimal
                              rewards:
                                type: number
                                format: decimal

  # Metrics and Monitoring
...